Move pylxd.exceptions to pylxd.deprecated.exceptions
