+++ pylxd/hosts.py
@@ -14,7 +14,7 @@
 from __future__ import print_function
 
 from pylxd.deprecated import base
-from pylxd import exceptions
+from pylxd.deprecated import exceptions
 
 
 class LXDHost(base.LXDBase):
