+++ pylxd/image.py
@@ -19,7 +19,7 @@ from six.moves import urllib
 
 from pylxd.deprecated import base
 from pylxd.deprecated import connection
-from pylxd import exceptions
+from pylxd.deprecated import exceptions
 from pylxd import mixin
 from pylxd.operation import Operation
