+++ pylxd/tests/test_host.py
@@ -17,7 +17,7 @@ from ddt import ddt
 import mock
 
 from pylxd.deprecated import connection
-from pylxd import exceptions
+from pylxd.deprecated import exceptions
 
 from pylxd.tests import annotated_data
 from pylxd.tests import fake_api
